+++ src/contrib/mlir/compiler/pass/mlir_subgraph_extraction.cpp
@@ -408,5 +408,5 @@ bool MLIRSubgraphExtractionPass::check_cycles(std::shared_ptr<Node> node,
 
 const std::set<std::type_index> MLIRSubgraphExtractionPass::m_supported_ops{
 #define MLIR_OP(OP) TI(ngraph::op::OP),
-#include "contrib/mlir/ops_supported.inc"
+#include "contrib/mlir/compiler/ops_supported.inc"
 };
