+++ src/ngraph/runtime/cpu/cpu_backend.cpp
@@ -27,7 +27,7 @@
 #include "ngraph/util.hpp"
 
 #ifdef NGRAPH_MLIR_ENABLE
-#include "contrib/mlir/compiler.hpp"
+#include "contrib/mlir/compiler/compiler.hpp"
 #endif
 
 using namespace ngraph;
